requirements.txt
# Shared / API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
redis>=5.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Ingest
paho-mqtt>=2.0.0

# Rules engine
celery>=5.3.0

# SCADA
opcua>=0.98.13
pymodbus>=3.5.0

# Dev
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
